requirements.txt
sounddevice
numpy
scipy
torch
transformers
openai-whisper
PyQt6
